Make rpidoor usuable for testing
We're a ways from a proper authentication system (see #7). In the meantime,
it's much more useful to be able to test the socket interface on something
that works. This is a low-importance device and easy to reconfigure one day.

In the meantime... Yes, I did just commit a password to a repository.