Working accessor example
`node --harmony create_modules.js` seems to work.. maybe? It's hard to be
certain when I can't post to a hue And seemingly simpler accessors all use
sockets which seemed like a lot of work to implement

All of this still needs to be tied to a server which creates modules out
of accessor json and then calls various methods when it gets POSTs and
GETs. That seems to be decently straightforward though