Snapshot of data collection for SWEC'15 paper
The goal was to compare runtime overhead for starting up the accessor
runtime (first run of an accessor) and subsequent runs. This was a
little rough, but certainly captures the spirit of what's going on. To
minimize the impact of I/O latency, a nop accessor that made http
requests to localhost was written (and the python server module loaded).

There was some funniness on my machine w.r.t. latency of 127.0.0.1 vs
localhost, but I took the faster of the two in each case.